requirements.txt
Django==2.0.5
django-filter==1.1.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
Markdown==2.6.11
Pillow==5.0.0
pytz==2018.4
six==1.11.0
